fix: tau is a float, not an int
Former-commit-id: 6812ff3d96a7711f2f8c45d8206160b5f980573a

+++ object-locator/argparser.py
@@ -279,7 +279,7 @@ def parse_command_args(training_or_testing):
             args.taus = [float(tau)
                          for tau in args.taus.replace('[', '').replace(']', '').split(',')]
         else:
-            args.taus = [int(args.taus)]
+            args.taus = [float(args.taus)]
 
         if isinstance(args.radii, (list, range)):
             pass
